Move metric configs to aggregator files and fix prediction_norm bug
Address PR #1123 review feedback:
- Move each metric config class to the file defining its aggregator,
  reducing main.py by ~300 lines
- Extract shared build infrastructure (MetricBuildContext,
  VariableFilterAdapter, etc.) into build_context.py
- Fix VariableFilterAdapter.record_batch() to guard prediction_norm
  access with has_prediction_norm check
- Move AnyAggregatorConfig type alias after all imports in train_config.py
