+++ fme/ace/aggregator/inference/__init__.py
@@ -1,21 +1,21 @@
+from ..one_step.ensemble import EnsembleMetricConfig
+from ..one_step.reduced import StepMeanMetricConfig
+from .annual import AnnualMetricConfig
+from .enso.dynamic_index import EnsoIndexMetricConfig
+from .enso.enso_coefficient import EnsoCoefficientMetricConfig
+from .histogram import HistogramMetricConfig
 from .main import (
-    AnnualMetricConfig,
-    EnsembleMetricConfig,
-    EnsoCoefficientMetricConfig,
-    EnsoIndexMetricConfig,
-    HistogramMetricConfig,
     InferenceAggregator,
     InferenceAggregatorConfig,
     InferenceEvaluatorAggregator,
     InferenceEvaluatorAggregatorConfig,
-    MeanMetricConfig,
     MetricConfig,
-    PowerSpectrumMetricConfig,
-    SeasonalMetricConfig,
     StepMeanEntry,
-    StepMeanMetricConfig,
-    TimeMeanMetricConfig,
     TypedMetricInferenceEvaluatorAggregatorConfig,
-    VideoMetricConfig,
-    ZonalMeanMetricConfig,
 )
+from .reduced import MeanMetricConfig
+from .seasonal import SeasonalMetricConfig
+from .spectrum import PowerSpectrumMetricConfig
+from .time_mean import TimeMeanMetricConfig
+from .video import VideoMetricConfig
+from .zonal_mean import ZonalMeanMetricConfig
