+++ fme/ace/aggregator/inference/annual.py
@@ -2,7 +2,7 @@ import dataclasses
 import datetime
 from collections.abc import Callable, Mapping
 from functools import partial
-from typing import Any, Optional
+from typing import Any, Literal, Optional
 
 import numpy as np
 import torch
@@ -16,7 +16,8 @@ from fme.core.gridded_ops import GriddedOperations
 from fme.core.typing_ import TensorMapping
 
 from ..plotting import plot_mean_and_samples
-from .data import InferenceBatchData
+from .build_context import MetricBuildContext, maybe_filter
+from .data import InferenceBatchData, SubAggregator
 
 
 class PairedGlobalMeanAnnualAggregator:
@@ -367,3 +368,27 @@ def to_dataset(data: TensorMapping, time: xr.DataArray) -> xr.Dataset:
 def _get_min_samples(timestep: datetime.timedelta) -> int:
     steps_per_day = datetime.timedelta(days=1) // timestep
     return 362 * steps_per_day
+
+
+@dataclasses.dataclass
+class AnnualMetricConfig:
+    type: Literal["annual"] = "annual"
+    variables: list[str] | None = None
+    name: str = "annual"
+    reference_data: str | None = None
+
+    def get_name(self) -> str:
+        return self.name
+
+    def build(self, ctx: MetricBuildContext) -> SubAggregator:
+        if self.reference_data is not None:
+            ref = xr.open_dataset(self.reference_data, decode_timedelta=False)
+        else:
+            ref = ctx.monthly_reference_data
+        agg: SubAggregator = PairedGlobalMeanAnnualAggregator(
+            ops=ctx.ops,
+            timestep=ctx.timestep,
+            variable_metadata=ctx.variable_metadata,
+            monthly_reference_data=ref,
+        )
+        return maybe_filter(agg, self.variables)
