+++ fme/ace/aggregator/inference/enso/__init__.py
@@ -1,6 +1,10 @@
 from .dynamic_index import (
+    EnsoIndexMetricConfig,
     LatLonRegion,
     PairedRegionalIndexAggregator,
     RegionalIndexAggregator,
 )
-from .enso_coefficient import EnsoCoefficientEvaluatorAggregator
+from .enso_coefficient import (
+    EnsoCoefficientEvaluatorAggregator,
+    EnsoCoefficientMetricConfig,
+)
